Add User-Agent to getRss
