+++ app/src/main/java/org/cssnr/tibs3dprints/CustomGlideModule.kt
@@ -1,7 +1,6 @@
 package org.cssnr.tibs3dprints
 
 import android.content.Context
-import android.util.Log
 import com.bumptech.glide.Glide
 import com.bumptech.glide.Registry
 import com.bumptech.glide.annotation.GlideModule
@@ -9,6 +8,7 @@ import com.bumptech.glide.integration.okhttp3.OkHttpUrlLoader
 import com.bumptech.glide.load.model.GlideUrl
 import com.bumptech.glide.module.AppGlideModule
 import okhttp3.OkHttpClient
+import org.cssnr.tibs3dprints.ui.news.getUserAgent
 import java.io.InputStream
 
 @GlideModule
@@ -25,14 +25,4 @@ class CustomGlideModule : AppGlideModule() {
         val factory = OkHttpUrlLoader.Factory(okHttpClient)
         registry.replace(GlideUrl::class.java, InputStream::class.java, factory)
     }
-}
-
-
-fun Context.getUserAgent(): String {
-    val versionName = this.packageManager.getPackageInfo(this.packageName, 0).versionName
-    val appName = this.getString(R.string.app_name)
-    val githubUrl = this.getString(R.string.website_url)
-    val userAgent = "${appName}/${versionName} - $githubUrl"
-    Log.d("getUserAgent", "userAgent: $userAgent")
-    return userAgent
 }
